replace withDashboardRouting with direct routing
